Declare array size member as size_t.
* src/filterrtf.cpp (RTFToken):
  - Declare RTFToken::valLen as size_t instead of int,
    fixing "signedness" compiler warning when comparing size.

+++ src/filterrtf.cpp
@@ -41,7 +41,7 @@ public:
 private:
   ETokenType type;
   char       val[1024];
-  int        valLen;
+  size_t     valLen;
 };
 
 inline void
